list categories needs further correction
